pyproject.toml
[project]
name = "network-security"
version = "0.0.1"
description = "Network Security MLOps Pipeline for Phishing Detection"
authors = [
    {name = "Abhishek Suresh", email = "as18757@nyu.edu"}
]
dependencies = [
    "python-dotenv",
    "pandas",
    "numpy",
    "pymongo[srv]>=4.0",
    "certifi",
    "scikit-learn",
    "scipy",
    "PyYAML",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["networksecurity"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8",
] 